중간

+++ back/controllers/likecommentController.js
@@ -1,32 +1,32 @@
 const { User, LikeComment, Comment } = require("../models");
 
-// 하트를 보여줄 수 있는 컨트롤러
+// // 하트를 보여줄 수 있는 컨트롤러
 
-exports.likeslist = async (req, res) => {
-  try {
-    const { id } = req.params;
+// exports.likeslist = async (req, res) => {
+//   try {
+//     const { id } = req.params;
 
-    const commentArr = await Comment.findAll({
-      where: { board_id: id }, include: {
-        model: LikeComment
-      }
-    })
-    const temp = commentArr.map((value) => {
-      return {
-        ...value, LikeComments: value.dataValues.LikeComments.map((value2) => {
-          return value2.dataValues
-        })
-      }
+//     const commentArr = await Comment.findAll({
+//       where: { board_id: id }, include: {
+//         model: LikeComment
+//       }
+//     })
+//     const temp = commentArr.map((value) => {
+//       return {
+//         ...value, LikeComments: value.dataValues.LikeComments.map((value2) => {
+//           return value2.dataValues
+//         })
+//       }
 
-    })
+//     })
 
-    // const data = await LikeComment.findAll()
-    res.json(data)
-  } catch (error) {
-    console.log("likeslist 오류터짐")
-    console.log(error)
-  }
-}
+//     // const data = await LikeComment.findAll()
+//     res.json(data)
+//   } catch (error) {
+//     console.log("likeslist 오류터짐")
+//     console.log(error)
+//   }
+// }
 // 하트가 true 면likeComment에 올라가는 컨트롤러
 exports.updateLikes = async (req, res) => {
   const comment_id = req.body.comment_id
@@ -47,13 +47,18 @@ exports.updateLikes = async (req, res) => {
 }
 // 하트 삭제 컨트롤러
 
-exports.deleteLikes = async () => {
-  const user_id = req.user.id;
-  const comment_id = req.params.comment_id;
+exports.deleteLikes = async (req, res) => {
+  const commentId = req.params.id
+  console.log("666666666",commentId)
+  const userData = req.decoded;
+  const UserId = userData.front_id;
   try {
-    await LikeComment.destroy({ where: { user_id, comment_id } });
+    const user = await User.findOne({ where: { user_id: UserId }});
+    const user_id = user.id;
+
+    await LikeComment.destroy({ where: { user_id: user_id,comment_id:commentId } });
     res.send("삭제됨?");
   } catch (error) {
-    console.log("삭제됨:", error);
+    console.log("삭제 에러:", error);
   }
 }
